# a4-auth/ai-rematch/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Supabase Auth API",
    "version": "1.0.0",
    "description": "Authentication API using Supabase Auth"
  },
  "servers": [{ "url": "http://localhost:3001" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/public/info": {
      "get": {
        "summary": "Public info",
        "responses": { "200": { "description": "Public message" } }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing fields" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tokens" },
          "400": { "description": "Missing fields" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}